uint -> long for Offset of Entry Add IsEncrypted to Entry by default Refactoring and comments Additional utility functions

+++ ArcFormats/Common/MultiFileArchive.cs
@@ -60,7 +60,7 @@ namespace GameRes.Formats
             }
         }
 
-        protected virtual uint GetEntrySize (Entry entry)
+        protected virtual long GetEntrySize (Entry entry)
         {
             return entry.Size;
         }
